+++ pycalphad/tests/test_equilibrium.py
@@ -10,10 +10,14 @@ from sympy import Symbol
 from numpy.testing import assert_allclose
 import numpy as np
 from pycalphad import Database, Model, calculate, equilibrium, EquilibriumError, ConditionError
-from pycalphad.codegen.callables import build_callables
+from pycalphad.codegen.callables import build_callables, build_phase_records
 from pycalphad.core.solver import SolverBase, Solver
+<<<<<<< HEAD
 from pycalphad.core.utils import get_state_variables
 from pycalphad.models.model_mqmqa import ModelMQMQA
+=======
+from pycalphad.core.utils import get_state_variables, instantiate_models
+>>>>>>> PyCalphad/master
 import pycalphad.variables as v
 from pycalphad.tests.datasets import *
 
@@ -57,6 +61,57 @@ def test_eq_binary():
     conds = {v.T: 1400, v.P: 101325, v.X('AL'): 0.55}
     eqx = equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True)
     assert_allclose(eqx.GM.values.flat[0], -9.608807e4)
+
+
+def test_phase_records_passed_to_equilibrium():
+    "Pre-built phase records can be passed to equilibrium."
+    my_phases = ['LIQUID', 'FCC_A1', 'HCP_A3', 'AL5FE2', 'AL2FE', 'AL13FE4', 'AL5FE4']
+    comps = ['AL', 'FE', 'VA']
+    conds = {v.T: 1400, v.P: 101325, v.N: 1.0, v.X('AL'): 0.55}
+
+    models = instantiate_models(ALFE_DBF, comps, my_phases)
+    phase_records = build_phase_records(ALFE_DBF, comps, my_phases, conds, models)
+
+    # With models passed
+    eqx = equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, model=models, phase_records=phase_records)
+    assert_allclose(eqx.GM.values.flat[0], -9.608807e4)
+
+
+def test_missing_models_with_phase_records_passed_to_equilibrium_raises():
+    "equilibrium should raise an error if all the active phases are not included in the phase_records"
+    my_phases = ['LIQUID', 'FCC_A1', 'HCP_A3', 'AL5FE2', 'AL2FE', 'AL13FE4', 'AL5FE4']
+    comps = ['AL', 'FE', 'VA']
+    conds = {v.T: 1400, v.P: 101325, v.N: 1.0, v.X('AL'): 0.55}
+
+    models = instantiate_models(ALFE_DBF, comps, my_phases)
+    phase_records = build_phase_records(ALFE_DBF, comps, my_phases, conds, models)
+
+    with pytest.raises(ValueError):
+        # model=models NOT passed
+        equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, phase_records=phase_records)
+
+
+def test_missing_phase_records_passed_to_equilibrium_raises():
+    "equilibrium should raise an error if all the active phases are not included in the phase_records"
+    my_phases = ['LIQUID', 'FCC_A1']
+    subset_phases = ['FCC_A1']
+    comps = ['AL', 'FE', 'VA']
+    conds = {v.T: 1400, v.P: 101325, v.N: 1.0, v.X('AL'): 0.55}
+
+    models = instantiate_models(ALFE_DBF, comps, my_phases)
+    phase_records = build_phase_records(ALFE_DBF, comps, my_phases, conds, models)
+
+    models_subset = instantiate_models(ALFE_DBF, comps, subset_phases)
+    phase_records_subset = build_phase_records(ALFE_DBF, comps, subset_phases, conds, models_subset)
+
+    # Under-specified models
+    with pytest.raises(ValueError):
+        equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, model=models_subset, phase_records=phase_records)
+
+    # Under-specified phase_records
+    with pytest.raises(ValueError):
+        equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, model=models, phase_records=phase_records_subset)
+
 
 @pytest.mark.solver
 def test_eq_single_phase():
@@ -596,6 +651,63 @@ def test_eq_issue259():
     assert_allclose(eq.GM.values, -65786.260)
     assert_allclose(eq.MU.values.flatten(), [-95906., -52877.592122])
 
+@pytest.mark.solver
+def test_eq_needs_metastable_starting():
+    """
+    Complex multi-component system with many phases near the starting hyperplane.
+    """
+    dbf = Database(MC_FECOCRNBTI_TDB)
+    phases = list(set(dbf.phases.keys()) - {'GP_MAT'})
+    mass_fracs = {v.W('CR'): 28./100, v.W('FE'): 21./100, v.W('NB'): 1./100, v.W('TI'): 1.3/100}
+    conds = v.get_mole_fractions(mass_fracs, 'CO', dbf)
+    conds[v.T] = 960
+    conds[v.P] = 1e5
+    conds[v.N] = 1
+    eq = equilibrium(dbf, ['CO', 'CR', 'FE', 'NB', 'TI', 'VA'], phases, conds)
+    assert set(np.squeeze(eq.Phase.values)) == {'SIGMA', 'BCC_A2', 'MU_PHASE', 'FCC_A1', 'LAVES_PHASE', ''}
+    assert_allclose(eq.GM.values, -46868.31620088)
+
+
+def test_eq_associate():
+    """
+    Associate model where the number of elements is different from the number of components (gh-367).
+    """
+    ASSOC_TDB = """
+    ELEMENT A BLANK 0.0 0.0 0.0 !
+    ELEMENT Q BLANK 0.0 0.0 0.0 !
+    SPECIES COMPA A1.0Q1.0 !
+    SPECIES COMPB Q2.0 !
+    SPECIES COMPC A1.0 !
+    SPECIES COMPD A1.5 !
+    TYPE_DEFINITION % SEQ * !
+    DEFINE_SYSTEM_DEFAULT ELEMENT 2 !
+    
+    PHASE PHASEA %  1 1.0 !
+    CONSTITUENT PHASEA :Q: !
+    
+    PHASE PHASEB %  1 1.0 !
+    CONSTITUENT PHASEB :Q: !
+    
+    PHASE PHASEC %  1 1.0 !
+    CONSTITUENT PHASEC :Q: !
+    
+    PHASE PHASED %  1 1.0 !
+    CONSTITUENT PHASED :Q: !
+    
+    PHASE PHASEE %  1 1.0 !
+    CONSTITUENT PHASEE :A: !
+    
+    PHASE PHASEF %  1 1.0 !
+    CONSTITUENT PHASEF :COMPB,COMPC: !
+    
+    PHASE PHASEG %  1 1.0 !
+    CONSTITUENT PHASEG :COMPA,COMPD: !
+    """
+    dbf = Database(ASSOC_TDB)
+    phases = list(set(dbf.phases.keys()))
+    conds = {v.X('Q'): 0.3, v.T: 500, v.P: 1e5, v.N: 1}
+    eq = equilibrium(dbf, ['A', 'Q'], phases, conds)
+    assert_allclose(eq.GM.values, -1736.981311)
 
 def test_MQMQA_equilibrium():